// apps/web/src/client/types/project-label.ts
export type ProjectLabel = {
  __typename: "ProjectLabel";
  id: string;
  projectId: string;
  name: string;
  color: string;
  createdAt: string;
  updatedAt: string;
};

export type ProjectLabelsResponse = {
  labels: ProjectLabel[];
};

export type ProjectLabelResponse = {
  label: ProjectLabel;
};